server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.3
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
paho-mqtt==1.6.0
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.5
ultralytics-thop==2.0.8
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
